fixed_free_list: - changed to set the last items link value to an invalid index - added an invalid link address to detect when the list is exhausted - added function to check if index is valid
sector_grid_index:
- removed combined version of the data. now relying on the grid helper to correctly get the sector value

sector_grid_helper:
- added function to get sector index from xy
- added function to get sector  and sub sector values

sector_grid_dimensions:
- added some extra helper values and converted uint32 size definition values to size_t as they should be compile time only

overlapTrackingGrid:
- applied refactor for the new way to access sector index and sub sector index

2d_physics_main:
- first pass at adding new object creation buffer with the goal of it being multi threadable maybe?

+++ sector_grid_data_structure/sector_grid_dimensions.h
@@ -6,9 +6,13 @@ namespace SectorGrid
 	/// <summary>
 	/// this struct is a helper stuct that defines the size of sectors in a sector grid along with the size of the grid
 	/// </summary>
-	template<uint32 TGridSectorW, uint32 TSectorTileW>
+	template<size_t TGridSectorW, size_t TSectorTileW>
 	struct sector_grid_dimensions
 	{
+		//using sector_value_type = MiscUtilities::uint_s<TGridSectorW * TGridSectorW>::int_type_t;
+		//using sub_sector_tile_value_type = MiscUtilities::uint_s<TSectorTileW * TSectorTileW>::int_type_t;
+		//using global_tile_value_type = MiscUtilities::uint_s<TSectorTileW * TSectorTileW * TGridSectorW * TGridSectorW>::int_type_t;
+
 		static constexpr uint32 sectors_grid_w = TGridSectorW;
 		static constexpr uint32 sector_grid_count = sectors_grid_w * sectors_grid_w;
